Move context into a separate module
This fixes an issue where bundlers can become confused causing the
context that's used when calling RelayRouter.Provider to be different
from the one called in RelayRouterLink. Moving the context to a separate
file ensures the code in the module is only executed once and all other
code uses the same value.