Force RelayRouter to have a single module entry point
This renames the modules in Relay Router to all be internal. The only
module that's kept is `RelayRouter` which exposes only those things that
are considered to be part of the API.

In this commit we focus on getting this set-up working. It also reveals
that more than the public API (e.g. __Bindings and __Internal) are
actually exposed because they're required by generated code.

We may want to move things that are needed in the generator to something
like RelayRouter.GeneratorHelper.* so it's clear for user code that
things like Bindings and Internal are not actually part of the API.