Improve contributor ergonomics
This restructures the most important scripts to make it easier to get
the project running locally. This previously required knowing exactly
which tools to use but now allows simply to run `yarn dev`.

We use `npm-run-all` to easily orchestrate our scripts.

Nodemon is added so that we can automatically reload our server in case
we make changes in `server.mjs`.

A `CONTRIBUTING.md` is created to explain the steps of getting the
project running locally with these new scripts. It can be expanded with
things like contributor agreements or other details.