Update to stable React 18.2.0
React 18.2.0 was released which is slightly newer than the "next"
version for the unreleased 18.2.0 preview we were previously using so we
update our main development dependencies.

We clear out the resolutions because we're happy with any 18.2.x version
we get. The property itself is left behind so we can easily switch back
to a preview version if needed.