WIP Make EntryServer a request handler
This is the start of converting the server.mjs to ReScript which
requires the addition of some bindings for server side code.

The main goal of these changes is to enable moving the React rendering
side and its effects (emission of assets) entirely into the EntryServer.

This should make it possible to create a pattern around renderTo*Stream
where utilities (e.g. the `PreloadInsertingStream` and others) can be
wired up to work regardless of the server environment.

Typing the original asset to script conversion code was too difficult to
complete in a day, so to ensure this commit brings code into a
buildable state that can be shared and interated upon.

There are two important TODO's that have significant changes from the
working code in another project:

1. Vite scripts are rewritten to script tags, in the other project these
are passed in as `head` in development mode (and head is empty in
production) but this makes EntryServer's default export bigger and
requires propagating this development snippet through the tree, which
feels wrong.

2. We write HTML to the stream manually. My suggestion is to make the
entire document part of the HTML tree. See the explanation and
motivation on line 73-100 in EntryServer.res.