Allow Vite plugin to transform arbitrary @rescriptModule references
Vite already helps us transform dynamic imports for our ReScript modules
(for which we don't know the name of the .mjs or .bs.js or .cjs file
when writing the ReScript code) to imports of the proper JS files.

However, in some places we also want to reference the file directly, for
example to be able to generate a `<script>` tag ourselves without use of
`import`. Up to now this relied on a runtime inspection of the ReScript
build artifacts.

This commit introduces a special non-existent function in ReScript which
serves as a marker for a newly introduced Vite plugin code transform.
This allows the code transform to replace the function invokation with
the literal string pointing to the absolute path of the source file.

This allows runtime code to use the ssrManifest generated by Vite to
turn this file-path into an asset chunk during production or to
transform it into a path relative to the Vite dev-server working
directory during development.