Remove tailwindcss from repository
While this is very useful in an actual application it's not used in our
development app. We can add it back later when we work on examples for
R3, but for now removing Tailwind helps clean up the repository.

This also cleans up PostCSS as it was used for Tailwind post-processing.