Move to React rendering the entire document
This commit replaces the `index.html` with a React component. This
allows React to have control over the entire document and makes it
possible to build functionality to swap out the head during client side
routing. Additionally this allows control over what part of the entire
document belongs to the React shell.

To ensure that we send a proper document to the browser we do not write
any assets before React has written an initial chunk (which should
include at least the doctype and html/head tags).

A ViteHead component is introduced which loads the fast-refresh scripts
during development. This replaces the transformIndexHtml call in
development.

We introduce a Vite plugin to ensure vite's compilation doesn't fail on
the absense of our `index.html` file. The file itself is removed to
ensure people don't make changes to it expecting something to happen.

The new plugin functionality is not merged into the existing plugin for
two reasons:

1) The decision on whether to use Vite's index.html or a React Html
component is opt-in for the Application (since it controls the server
and rendering, it could also load and transform a static index.html).
Splitting this ensures that decision remains with the developer.

2) The plugin must run `pre` to ensure it runs before the Vite core html
build plugin. The primary plugin has no such restriction so keeeping it
separate makes sense.