Move PreloadInsertingStream into router package
This allows applications to use the stream to easily inject assets
rather than having to provide their own.

This chooses to use an opaque type for the stream so that the router
does not need Node.js bindings, even though this makes consumption
slightly more difficult.

The name uses the `Node` suffix to indicate this is Node.js specific. We
can create a parallel `PreloadInsertingStreamTransform` for platforms
that use native web streams.

Fixes #66