Move non-optional parameters after optional parameters
This makes it easier to pipe into functions or do partial application of
functions.

This also introduces types for the preload functions so that they're
consistent throughout the code-base.

We rename `makeClientAssetPreloader` to better indicate that it takes
only a single argument and returns a new function that you should give
to something else. With than in mind we also make it an unnamed
argument. Unfortunately ReScript format puts together the argument
declaration which makes it less obvious in code that it returns a
function, though its interface signature makes that clear now.

Closes #25