Ensure entry point direct imports are part of bootstrapModules
The `bootstrapModules` for React indicate the scripts that will be
needed to start hydrating, so React adds them to the page for us as soon
as possible after the shell is produced. That allows the user to see the
initial HTML as soon as possible and then start loading what's needed
for hydration to start.

Previously we only included our compiled EntryClient and we omitted any
static imports that that may have. This commit ensures that any file we
know will be required as soon as our EntryClient is loaded are just sent
to the browser immediately, this reduces the waterfall effect for
initial script loading.