Change script insertion to `import` for immediate preloads
We've been informed by trustable sources that `import` is equal to a
dynamically inserted `<script type="module">` (and that `async` is
implied when dynamically inserting scripts).

To simplify our code, for the scripts we want to preload immediately
rather than whenever the browser has time we use an asymc import.