Move to monorepo structure
This moves all code into a monorepo structure. It splits up the packages
that make-up and support the router into isolated units. We split the
two Vite plugins since they may also be useful outside the router.
Similarly we split the CLI from the router to indicate that it is indeed
a separate entity.

We move the Vitest bindings to a shared package since more than one of
our internal packages use them. For now we mark this package as private,
though the fact that our `example` repo also depends on it may mean we
want to publish it in the future.

We add yarn workspace tools to help us manager our monorepos.

package.json and bsconfig.json files are updated for each individual
package and example to clearly specify their own dependencies and ensure
ReScript can find what it needs.

The top level package.json only gets scripts that delegate to the
individual packages. We let yarn figure out the topology to build any
dependants in order.

We fix the Vite plugin test while we're at it.

The code from `buildPackage.sh` that was used to add a shebang to our
Cli is moved to the CLI package.json' build script instead. We can now
remove the buildPackage.sh since we can create publish scripts for the
individual packages without need for other modifications.