Configure .npmignore files for packages
We don't want to publish everything in every package to NPM. For the
plugins and the CLI only the JavaScript files are needed. For the router
we expect people to use the package in their ReScript project so we
don't ship any compiled files.

This matches what was previously included manually in the `_release`
package.

It does fix a bug where `PreloadInsertingStream.mjs` was not included in
releases.