Set-up changesets
Changesets allows us to manage the versioning of the packages in our
repository and can help us with releasing.

Additionally changeset files allow us to describe changes and their
impact on package versions to influence how versioning will happen and
keep track of a changelog.

A few configuration changes have been made with respect to the default
configuration. The docs for config can be found at
https://github.com/changesets/changesets/blob/main/docs/config-file-options.md

- The main router package and cli will always have the same version
  (configured as fixed) even if only one of the two has changes
- `access` has been set to `public` since we publish public packages
- `updateInternalDependencies` has been set to `minor` which provides
  package consumers a bit more room on what package versions they use
  and does not strictly require the use of the latest patch version.
  This was recommended by the documentation.