Add automated publishing workflow
This adds a release workflow. The workflow is based on
https://github.com/changesets/action#with-publishing

To complete setup we must add a `GITHUB_TOKEN` and `NPM_TOKEN` (without
2FA for publishing) to the GitHub repository secrets.

The workflow will create a pull request when new changesets are found on
the main branch. When this pull request is merged the versions of
packages will be updated and a new release on NPM will be created
automatically. When more changes are added to the main branch the pull
request will automatically be updated.