Transform Vite's client manifest into a Router manifest
The format of Vite's client manifest is not entirely useeful to us and
contains a bit more information than we need, this makes it impractical
to ship to the client.

We provide a type definition for a new manifest structure that easily
gives us insights into what we should load (imports, css files and
assets) for a specific JavaScript file. Additionally we provide some
tooling to easily convert a Vite manifest to this new strucutre.