Namespace packages and create primary package
This moves the router specific packages under the
`@rescript-relay-router` namespace. The vite plugin package is renamed
to `/vite` with the idea that it can contain everything that's specific
to supporting Vite with the router (so we could eventually also move
`Vite` bindings to this package).

The virtual html Vite plugin is renamed to follow Vite plugin naming
conventions of starting with `vite-plugin`. The name is changed to
`virtual-index` to not clash with existing `virtual-html` plugins. This
means it can be published to NPM separately and used by others.

We move `router` to `core` and introduce a new `router` package that can
be published under `rescript-relay-router`. Unfortunately the ReScript
compiler doesn't do well with these meta packages so we must still add
`@rescript-relay-router/core` to our `express` example's
`bsconfig.json`. The specific issue is that ReScript doesn't let
`express` use dependencies of `router` without specifying them
themselves.