Pass manifest to EntryServer.handleRequest
Rather than providing the bootstrapModules for React we now provide the
entryPoint JavaScript file as well as the RelayRouter manifest. This
allows us to change our preloadAsset function to also preload any other
files the component may need.

In addition we can now determine the bootstrapModules in our EntryServer
and trigger preloads for any css and other assets that our initial
scripts may need.

We could also decide not to use React's `bootstrapModules` here and
instead just trigger a `preloadAsset` on `entryPoint` which would also
take care of preloading `image` and `css` files for us.