Write router manifest to disk and transform in Vite plugin
This commit moves the transform from the Vite manifest file to our
specialised router manifest into the Vite plugin rather than in the
server code. This means the server only has to load the file.

We change the manifest slightly to also include the main entry point for
our application, this ensures the server doesn't also need to look at
Vite's manifest to figure out the entry point. It reduces the amount of
arguments to the requestHandler too.

I've chosen to read and write Vite's `manifest.json` at the end of the
plugin rather than duplicating Vite's manifest plugin efforts. We can
choose to duplicate Vite's manifest plugin implementation directly so we
can actually emit the file and have it reported in the future, though my
advice would be to ensure the functionality itself then lives in an
isolated class.