Move History module to its own first-class module file
Publishing the History module as separate package makes it too difficult
to install for now (until we have e.g.
`yarn create rescript-relay-router-app` support).

To ensure that we can keep code simple and move History out of the
package in the future without breaking change we do promote it to a
first class module that's no longer exposed under the RelayRouter
namespace.

This also makes the History module as-is more accessible to people who
want to contribute extensions to be able to manipulate history data
outside of the router.