Add integration tests folder
We want to test some cases for which traditional testing tools are not
entirely suitable. For example testing that a `routes.json` file turns
into compilable ReScript code. For this we create workspaces in
`tests/*` that allow us to test specific cases.

The initial added case demonstrates an issue with using reserved words
in parameter names or query parameter names.