Namespace History into RelayRouter__History
We moved it to a separate module and a global namespace. The idea was
that we could move it to a standalone package without making it a
breaking change. However, that appears not to be possible anyway so it
makes more sense for now to ensure the history module is in the
RelayRouter namespace to avoid any potential conflicts.