Ensure RescriptRelayRouterCli.mjs is executable
This adds a post-build script to ReScript that prepends a shebang line
to the compiled ReScriptRelayRouterCli.mjs file. The overhead of a bash
script should be neglible compared to the ReScript compiler.

Using a post-build hook rather than something external to ReScript
ensures that the shebang is only ever added when the file is
regenerated (avoiding duplication) and is also prepended when ReScript
is run in watch mode.